Compatible with accelerate==0.21.*
